feature(287): Added ability to convert Unicode decimal digits to digits specified by a target zero-digit.
This means that for custom types that accept numeric values, the user or client could provide them using the decimal digits most familiar to them.

This also caters to users copying-and-pasting numeric values from one screen to another where the copy-from is presented in, say, Devanagari digits, and the target application being pasted-to is expecting Western Arabic digits.